+++ kvm-ioctls/src/cap.rs
@@ -41,7 +41,12 @@ pub enum Cap {
     Iommu = KVM_CAP_IOMMU,
     DestroyMemoryRegionWorks = KVM_CAP_DESTROY_MEMORY_REGION_WORKS,
     UserNmi = KVM_CAP_USER_NMI,
-    #[cfg(any(target_arch = "x86_64", target_arch = "aarch64", target_arch = "s390x"))]
+    #[cfg(any(
+        target_arch = "x86_64",
+        target_arch = "aarch64",
+        target_arch = "s390x",
+        target_arch = "loongarch64"
+    ))]
     SetGuestDebug = KVM_CAP_SET_GUEST_DEBUG,
     #[cfg(target_arch = "x86_64")]
     ReinjectControl = KVM_CAP_REINJECT_CONTROL,