+++ kvm-ioctls/src/ioctls/vm.rs
@@ -573,7 +573,8 @@ impl VmFd {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     pub fn signal_msi(&self, msi: kvm_msi) -> Result<c_int> {
         // SAFETY: Safe because we allocated the structure and we know the kernel
@@ -622,13 +623,22 @@ impl VmFd {
     /// })
     /// .expect("Cannot create KVM vAIA device.");
     ///
+    /// #[cfg(target_arch = "loongarch64")]
+    /// vm.create_device(&mut kvm_bindings::kvm_create_device {
+    ///     type_: kvm_bindings::kvm_device_type_KVM_DEV_TYPE_LOONGARCH_EIOINTC,
+    ///     fd: 0,
+    ///     flags: 0,
+    /// })
+    /// .expect("Cannot create KVM EIOINTC device.");
+    ///
     /// let irq_routing = kvm_irq_routing::default();
     /// vm.set_gsi_routing(&irq_routing).unwrap();
     /// ```
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     pub fn set_gsi_routing(&self, irq_routing: &kvm_irq_routing) -> Result<()> {
         // SAFETY: Safe because we allocated the structure and we know the kernel
@@ -811,7 +821,7 @@ impl VmFd {
     /// # let vm = kvm.create_vm().unwrap();
     /// // This example is based on https://lwn.net/Articles/658511/.
     /// let mem_size = 0x4000;
-    /// let guest_addr: u64 = 0x1000;
+    /// let guest_addr: u64 = 0x4000;
     /// let load_addr: *mut u8 = unsafe {
     ///     libc::mmap(
     ///         null_mut(),
@@ -850,6 +860,13 @@ impl VmFd {
     ///     0xa3, 0x23, 0x73, 0x00, // sw t2, t1 + 7;   dirty current page
     ///     0x23, 0x20, 0x75, 0x00, // sw t2, a0;       trigger MMIO exit
     ///     0x6f, 0x00, 0x00, 0x00, // j .;shouldn't get here, but if so loop forever
+    /// ];
+    /// #[cfg(target_arch = "loongarch64")]
+    /// let asm_code = [
+    ///     0x0d, 0x00, 0x00, 0x18, // pcaddi t1, 0;    <this address> -> t1
+    ///     0xae, 0x1d, 0x80, 0x29, // st.w t2, t1, 7;  dirty current page
+    ///     0x8e, 0x00, 0x80, 0x29, // st.w t2, a0, 0;  trigger MMIO exit
+    ///     0x00, 0x00, 0x00, 0x50, // b .;shouldn't get here, but if so loop forever
     /// ];
     ///
     /// // Write the code in the guest memory. This will generate a dirty page.
@@ -896,6 +913,14 @@ impl VmFd {
     ///     let mmio_addr: u64 = guest_addr + mem_size as u64;
     ///     vcpu_fd.set_one_reg(core_reg_base, &guest_addr.to_le_bytes()); // set PC
     ///     vcpu_fd.set_one_reg(core_reg_base + 10, &mmio_addr.to_le_bytes()); // set A0
+    /// }
+    ///
+    /// #[cfg(target_arch = "loongarch64")]
+    /// {
+    ///     let mut vcpu_regs = vcpu_fd.get_regs().unwrap();
+    ///     vcpu_regs.pc = guest_addr;
+    ///     vcpu_regs.gpr[4] = guest_addr + mem_size as u64;
+    ///     vcpu_fd.set_regs(&vcpu_regs).unwrap();
     /// }
     ///
     /// loop {
@@ -975,7 +1000,8 @@ impl VmFd {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     pub fn register_irqfd(&self, fd: &EventFd, gsi: u32) -> Result<()> {
         let irqfd = kvm_irqfd {
@@ -1026,7 +1052,8 @@ impl VmFd {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     pub fn register_irqfd_with_resample(
         &self,
@@ -1084,7 +1111,8 @@ impl VmFd {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     pub fn unregister_irqfd(&self, fd: &EventFd, gsi: u32) -> Result<()> {
         let irqfd = kvm_irqfd {
@@ -1154,7 +1182,8 @@ impl VmFd {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     pub fn set_irq_line(&self, irq: u32, active: bool) -> Result<()> {
         let mut irq_level = kvm_irq_level::default();
@@ -1280,6 +1309,8 @@ impl VmFd {
     ///     type_: kvm_device_type_KVM_DEV_TYPE_ARM_VGIC_V3,
     ///     #[cfg(target_arch = "riscv64")]
     ///     type_: kvm_device_type_KVM_DEV_TYPE_RISCV_AIA,
+    ///     #[cfg(target_arch = "loongarch64")]
+    ///     type_: kvm_bindings::kvm_device_type_KVM_DEV_TYPE_LOONGARCH_IPI,
     ///     fd: 0,
     ///     flags: KVM_CREATE_DEVICE_TEST,
     /// };
@@ -1296,6 +1327,8 @@ impl VmFd {
     ///     }
     ///     #[cfg(target_arch = "riscv64")]
     ///     panic!("Cannot create vAIA device.");
+    ///     #[cfg(target_arch = "loongarch64")]
+    ///     panic!("Cannot create IPI device.");
     /// });
     /// ```
     pub fn create_device(&self, device: &mut kvm_create_device) -> Result<DeviceFd> {
@@ -2040,6 +2073,60 @@ pub(crate) fn request_aia_init(vaia: &DeviceFd) {
     vaia.set_device_attr(&vaia_attr).unwrap();
 }
 
+/// Create a dummy IPI device.
+///
+/// # Arguments
+///
+/// * `vm` - The vm file descriptor.
+/// * `flags` - Flags to be passed to `KVM_CREATE_DEVICE`.
+#[cfg(test)]
+#[cfg(target_arch = "loongarch64")]
+pub(crate) fn create_ipi_device(vm: &VmFd, flags: u32) -> DeviceFd {
+    let mut ipi_device = kvm_create_device {
+        type_: kvm_device_type_KVM_DEV_TYPE_LOONGARCH_IPI,
+        fd: 0,
+        flags,
+    };
+    vm.create_device(&mut ipi_device)
+        .expect("Cannot create KVM IPI device")
+}
+
+/// Create a dummy EIOINTC device.
+///
+/// # Arguments
+///
+/// * `vm` - The vm file descriptor.
+/// * `flags` - Flags to be passed to `KVM_CREATE_DEVICE`.
+#[cfg(test)]
+#[cfg(target_arch = "loongarch64")]
+pub(crate) fn create_eiointc_device(vm: &VmFd, flags: u32) -> DeviceFd {
+    let mut eiointc_device = kvm_create_device {
+        type_: kvm_device_type_KVM_DEV_TYPE_LOONGARCH_EIOINTC,
+        fd: 0,
+        flags,
+    };
+    vm.create_device(&mut eiointc_device)
+        .expect("Cannot create KVM EIOINTC device")
+}
+
+/// Create a dummy PCHPIC device.
+///
+/// # Arguments
+///
+/// * `vm` - The vm file descriptor.
+/// * `flags` - Flags to be passed to `KVM_CREATE_DEVICE`.
+#[cfg(test)]
+#[cfg(target_arch = "loongarch64")]
+pub(crate) fn create_pchpic_device(vm: &VmFd, flags: u32) -> DeviceFd {
+    let mut pchpic_device = kvm_create_device {
+        type_: kvm_device_type_KVM_DEV_TYPE_LOONGARCH_PCHPIC,
+        fd: 0,
+        flags,
+    };
+    vm.create_device(&mut pchpic_device)
+        .expect("Cannot create KVM PCHPIC device")
+}
+
 #[cfg(test)]
 mod tests {
     #![allow(clippy::undocumented_unsafe_blocks)]
@@ -2411,6 +2498,36 @@ mod tests {
     }
 
     #[test]
+    #[cfg(target_arch = "loongarch64")]
+    fn test_register_unregister_irqfd() {
+        let kvm = Kvm::new().unwrap();
+        let vm_fd = kvm.create_vm().unwrap();
+        let evtfd1 = EventFd::new(EFD_NONBLOCK).unwrap();
+        let evtfd2 = EventFd::new(EFD_NONBLOCK).unwrap();
+        let evtfd3 = EventFd::new(EFD_NONBLOCK).unwrap();
+
+        // Create the interrupt devices.
+        create_ipi_device(&vm_fd, 0);
+        create_eiointc_device(&vm_fd, 0);
+        create_pchpic_device(&vm_fd, 0);
+        vm_fd.create_vcpu(0).unwrap();
+
+        vm_fd.register_irqfd(&evtfd1, 4).unwrap();
+        vm_fd.register_irqfd(&evtfd2, 8).unwrap();
+        vm_fd.register_irqfd(&evtfd3, 4).unwrap();
+        vm_fd.unregister_irqfd(&evtfd2, 8).unwrap();
+        // KVM irqfd doesn't report failure on this case:(
+        vm_fd.unregister_irqfd(&evtfd2, 8).unwrap();
+
+        // Duplicated eventfd registration.
+        // On loongarch64 this fails as the event fd was already matched with a GSI.
+        vm_fd.register_irqfd(&evtfd3, 4).unwrap_err();
+        vm_fd.register_irqfd(&evtfd3, 5).unwrap_err();
+        // KVM irqfd doesn't report failure on this case:(
+        vm_fd.unregister_irqfd(&evtfd3, 5).unwrap();
+    }
+
+    #[test]
     #[cfg(target_arch = "x86_64")]
     fn test_set_irq_line() {
         let kvm = Kvm::new().unwrap();
@@ -2491,6 +2608,23 @@ mod tests {
 
         // Initialize valid vAIA device.
         request_aia_init(&vaia_fd);
+
+        vm_fd.set_irq_line(7, true).unwrap();
+        vm_fd.set_irq_line(7, false).unwrap();
+        vm_fd.set_irq_line(7, true).unwrap();
+    }
+
+    #[test]
+    #[cfg(target_arch = "loongarch64")]
+    fn test_set_irq_line() {
+        let kvm = Kvm::new().unwrap();
+        let vm_fd = kvm.create_vm().unwrap();
+        vm_fd.create_vcpu(0).unwrap();
+
+        // Create the interrupt devices.
+        create_ipi_device(&vm_fd, 0);
+        create_eiointc_device(&vm_fd, 0);
+        create_pchpic_device(&vm_fd, 0);
 
         vm_fd.set_irq_line(7, true).unwrap();
         vm_fd.set_irq_line(7, false).unwrap();
@@ -2616,7 +2750,8 @@ mod tests {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     fn test_signal_msi_failure() {
         let kvm = Kvm::new().unwrap();
@@ -2626,7 +2761,11 @@ mod tests {
     }
 
     #[test]
-    #[cfg(not(any(target_arch = "aarch64", target_arch = "riscv64")))]
+    #[cfg(not(any(
+        target_arch = "aarch64",
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
+    )))]
     fn test_enable_cap_failure() {
         let kvm = Kvm::new().unwrap();
         let vm = kvm.create_vm().unwrap();
@@ -2664,7 +2803,8 @@ mod tests {
     #[cfg(any(
         target_arch = "x86_64",
         target_arch = "aarch64",
-        target_arch = "riscv64"
+        target_arch = "riscv64",
+        target_arch = "loongarch64"
     ))]
     fn test_set_gsi_routing() {
         let kvm = Kvm::new().unwrap();
@@ -2681,6 +2821,11 @@ mod tests {
         // committing irq_routing table.
         #[cfg(target_arch = "riscv64")]
         create_aia_device(&vm, 0);
+
+        // LoongArch 64-bit expect an EIOINTC device to be created in advance of
+        // committing irq_routing table.
+        #[cfg(target_arch = "loongarch64")]
+        create_eiointc_device(&vm, 0);
 
         vm.set_gsi_routing(&irq_routing).unwrap();
     }